+++ libc/src/signal/linux/sigaltstack.cpp
@@ -34,7 +34,8 @@ LLVM_LIBC_FUNCTION(int, sigaltstack,
     }
   }
 
-  int ret = __llvm_libc::syscall_impl(SYS_sigaltstack, ss, oss);
+  int ret =
+      static_cast<int>(__llvm_libc::syscall_impl(SYS_sigaltstack, ss, oss));
   if (ret < 0) {
     libc_errno = -ret;
     return -1;
